Created ui.js and extracted functions from editor.js there
